Update Calculator.cpp
Executive Summary:
The goal of this project is to create a simple calculator that can calculate infix and postfix
expression. In this project, the calculator must use infix and postfix notation. I've used
object-oriented programming language for this project. By this project, I've ensured all
project requirements are met and all results are correct. I've completed the solution of
this project step-by-step. After running the code, you'll have two options - infix and postfix.
Now if you choose the infix option then you've to write an infix expression. If you choose the
other, you've to write the post fix expression. And finally, you'll see your final result
through Inter Press. This is how this calculator will work. This program has six-featured
function like infix, postfix, addition, subtraction, multiplication and division. I've tried
to keep the time complexity low so that the calculator can work efficiently and accurately.

In an infix expression, the operators stand between the operands, and in postfix, expression
operands stand before their operators. The operators precede their operands. In infix
expression the operators stand between the operands; on the other hand, in postfix the
operands follow the operators to reach the result. It is also known as the “Reverse Polish
Notation” which is evaluated easily and efficiently by computer but it is hard to read by
humans. For example, AB+ is reverse Polish for A+B. So, the purpose of this project is to
make a calculator to calculate the infix and postfix expressions. Generally, infix expression is
used to solve any mathematical problem, but a computer could not do it, it has to use prefix
or postfix expression, which it does automatically. In this project, I did the same thing but
by developing a program manually. I'll have to design and implement a program that will
work as a calculator. Using stack for calculating infix and postfix expressions.